requirements.txt
python-docx == 1.1.2
lxml >= 2.3.2
packaging == 24.1
requests
filelock == 3.15.4